Move source code to root
