+++ file_manager.py
@@ -1,0 +1,25 @@
+import time
+from watchdog.observers import Observer
+from watchdog.events import LoggingEventHandler
+
+
+class NewFileHandler(FileSystemEventHandler):
+    def on_created(self, event):
+        if event.is_directory:
+            return
+
+        print(f"New file created: {event.src_path}")
+
+
+def watch_for_new_files(path):
+    event_handler = NewFileHandler()
+    observer = Observer()
+    observer.schedule(event_handler, path, recursive=True)
+    observer.start()
+
+    try:
+        while True:
+            time.sleep(1)
+    except KeyboardInterrupt:
+        observer.stop()
+    observer.join()
