+++ tests/bad_config.json
